Added --seconds parameter, more statistics
Lets "--seconds" be an argument to the program, to limit the run based
on time. If both --iterations and --seconds are defined, the larger
limit is the deciding one.

Also added new tables to the result database, which track the voting
history of the games, as well as the mission results.